+++ projects/cm-d3-graphs/src/lib/interfaces/graph-configuration.interface.ts
@@ -154,6 +154,7 @@ export interface FlowChartGraphConfigurationInterface extends BaseGraphConfigura
     distanceBetweenParentAndChild?: number;
     expandable?: boolean;
     maxInitialExpandedLevel?: number;
+    icon?: string;
   };
   links?: {
     color?: string;