+++ src/app/mocks/flow-chart-confs.ts
@@ -4,7 +4,9 @@ export const FlowChartConfs: FlowChartGraphConfigurationInterface = {
   id: 'flow_chart',
   type: 'flow-chart',
   nodes: {
-    shape: 'rect'
+    shape: 'circle',
+    circleRadius: 50,
+    icon: 'https://upload.wikimedia.org/wikipedia/commons/thumb/1/12/User_icon_2.svg/2048px-User_icon_2.svg.png'
   },
   zoom: {
     minZoom: 0.5,