+++ RELEASENOTES.md
@@ -12,3 +12,10 @@
 | -------- | ---- | ---- | ---- |
 | Second release | **04-02-2021** | NEW | Updated readme |
 
+## [0.0.3]
+
+| Activity | Date | Type | Notes |
+| -------- | ---- | ---- | ---- |
+| Flow chart cluster position | **15-09-2021** | BUG FIX | Resolved a bug when a full space cluster is used |
+| Flow chart node icon | **15-09-2021** | IMPROVEMENT | Added icon to nodes |
+